requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
openai-whisper==20231117
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
sqlite3
requests==2.31.0
pydantic==2.5.0
elevenlabs==0.2.26
websockets==12.0
aiofiles==23.2.1
